v1: round 13: repay judged before the pull; outputs valued before the route; reviewed tokens only
Fixes for the reviewer's round 11/12 pass (FLIP-280, G11-H1, G12-H1,
G12-H2, G11-L1). Their 16 tests are imported; the four gap tests are
now named refusals.

G11-H1: AaveV3AdapterV1.snapshot returns the owner's health factor
before the core pulls anything (it returned nothing), and repay from
collateral gates on it (at or over the target: refused) and requires the
position after the firing to beat it. The pull can no longer open the
gate or count as progress. The unused fee reserve and unsold collateral
come back after the check, so the final health factor is higher still.
Tests: an owner already over the target is refused with no fee (fork);
a step that beats the post-pull position but not the pre-firing one is
refused (the reviewer's 95 % threshold model).

G12-H1: one valuation for every oracle transform, one output or
several: each output's and the asset's value per unit (price times WAD
over 10^decimals) is read before the route, and the same numbers judge
what arrived, so a route cannot move the prices it is judged by. The
one-output oracle check now shares this path (its OutputBelowMinimum
numbers are in value units); Fixed, Floor and ERC-4626 keep theirs. The
duplicate price math went, so GenericExecutorV1 is 24,382 bytes (194
spare) with every public getter as in round 12.

G12-H2: every output of a several-output transform must be a token the
registry bound a price round for (reviewed, independent assets), so two
addresses over one balance cannot be counted twice unless the admin
listed both. The app keeps the same list (the six X Layer feeds).

G11-L1: the scaled-token fuzz seeds are uint256 and span 1x to 2x.

Tests updated: several-output tests bind a mock Chainlink round for
their tokens, as the deploy does; USDG cannot be one of several outputs
but a one-token oracle swap into it still works. 474 tests, lint and
format clean.

+++ test/v1/ShieldV1.t.sol
@@ -142,10 +142,10 @@ contract ShieldV1Test is Test {
     function testFuzz_fixFeeRefundSurvivesScaledTokenRounding(
         uint96 amountSeed,
         uint16 spendSeed,
-        uint64 indexFrac
+        uint256 indexFrac
     ) public {
         // Aave's liquidity index is at least one ray and grows slowly; up to 2x here.
-        (MockScaledToken a, bytes32 id) = _scaledMandate(1e27 + bound(uint256(indexFrac), 0, 1e27));
+        (MockScaledToken a, bytes32 id) = _scaledMandate(1e27 + bound(indexFrac, 0, 1e27));
         uint256 amount = bound(amountSeed, 1e12, 100e18);
         exec.setSpendBps(bound(spendSeed, 1, 9_999)); // part unspent: the unused fee reserve goes back
         vm.prank(agent);
@@ -159,9 +159,9 @@ contract ShieldV1Test is Test {
     function testFuzz_scaledTokenChargeMatchesTheOwnersOutflow(
         uint96 amountSeed,
         uint16 spendSeed,
-        uint64 indexFrac
+        uint256 indexFrac
     ) public {
-        (MockScaledToken a, bytes32 id) = _scaledMandate(1e27 + bound(uint256(indexFrac), 0, 1e27));
+        (MockScaledToken a, bytes32 id) = _scaledMandate(1e27 + bound(indexFrac, 0, 1e27));
         uint256 amount = bound(amountSeed, 1e12, 100e18);
         exec.setSpendBps(bound(spendSeed, 1, 9_999));
         uint256 ownerBefore = a.balanceOf(principal);
